requirements.txt
PyYAML
behave
boto3
fastapi
jsonschema
kafka-python
minio
numpy
pandas
psycopg2-binary
pyarrow
pytest
python-multipart
requests
semver
sentry-sdk
toml
types-PyYAML
types-jsonschema
types-psycopg2
types-requests
uvicorn[standard]
virtualenv
setuptools>=70.0.0